requirements.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.2
tzdata==2022.2
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.13.0